+++ Sources/BrightroomUI/Built-in UI/ClassicImageEdit/ClassicImageEditOptions.swift
