+++ Sources/BrightroomUI/Built-in UI/ClassicImageEdit/ClassicImageEditViewModel.swift
