+++ Sources/BrightroomUI/Built-in UI/ClassicImageEdit/Components/ClassicImageEditControlStackView.swift
