+++ Sources/BrightroomUI/Built-in UI/ClassicImageEdit/Components/ClassicImageEditControlViewBase.swift
