+++ Sources/BrightroomUI/Built-in UI/ClassicImageEdit/Components/ClassicImageEditNavigationView.swift
@@ -45,8 +45,8 @@ open class ClassicImageEditNavigationView : UIStackView {
     cancelButton.setTitle(L10n.cancel, for: .normal)
     saveButton.setTitle(L10n.done, for: .normal)
 
-    cancelButton.setTitleColor(Style.default.black, for: .normal)
-    saveButton.setTitleColor(Style.default.black, for: .normal)
+    cancelButton.setTitleColor(ClassicImageEditStyle.default.black, for: .normal)
+    saveButton.setTitleColor(ClassicImageEditStyle.default.black, for: .normal)
 
     cancelButton.titleLabel!.font = UIFont.systemFont(ofSize: 17)
     saveButton.titleLabel!.font = UIFont.boldSystemFont(ofSize: 17)
