+++ Sources/BrightroomUI/Built-in UI/ClassicImageEdit/L10n.swift
