+++ Sources/BrightroomUI/Shared/Components/Crop/CropView.CropInsideOverlay.swift
