+++ Sources/BrightroomUI/Shared/Components/Crop/CropView.CropOutsideOverlay.swift
