+++ Sources/BrightroomUI/Shared/Components/Crop/CropView._CropScrollView.swift
