+++ Sources/BrightroomUI/Shared/Components/Crop/CropView._InteractiveCropGuideView.swift
