+++ Sources/BrightroomUI/Shared/Components/Crop/CropView.swift
