+++ Sources/BrightroomUI/Shared/Components/Crop/SwiftUICropView.swift
