+++ Sources/BrightroomUI/Shared/Components/Drawing/BlurryMaskingView.swift
