+++ Sources/BrightroomUI/Shared/Components/Drawing/CanvasView.swift
