+++ Sources/BrightroomUI/Shared/Components/Drawing/SmoothPathDrawingView.swift
