+++ Sources/BrightroomUI/Shared/Components/ImageViews/CIImageDisplaying.swift
