+++ Sources/BrightroomUI/Shared/Components/ImageViews/ImagePreviewView.swift
