+++ Sources/BrightroomUI/Shared/Components/ImageViews/MetalImageView.swift
