+++ Sources/BrightroomUI/Shared/Utils/EditingCrop+.swift
